changes made according to PR review
