+++ app/src/main/java/com/github/se/polyfit/di/UserModule.kt
@@ -30,14 +30,6 @@ object UserModule {
     // firestoredatabase needs a user id to be initialized
     return User(id = "testUserID")
   }
-  /*
-  @Provides
-  @Singleton
-  fun providesGraphViewModel(dataProcessor: LocalDataProcessor): GraphViewModel {
-    return GraphViewModel(dataProcessor)
-  }
-
-   */
 
   @Provides
   @Singleton
