+++ app/src/main/java/com/github/se/polyfit/ui/navigation/Navigation.kt
@@ -8,6 +8,10 @@ class Navigation(private val navHostController: NavHostController) {
     navHostController.popBackStack()
   }
 
+  fun navigateToGraph() {
+    navigateTo(Route.Graph)
+  }
+
   fun navigateToHome() {
     navigateTo(Route.Home)
   }
